Rename jobs for Asset Store Controller Manager
