+++ prow/jobs/kyma/components/asset-store-controller-manager/asset-store-controller-manager.yaml
@@ -16,6 +16,25 @@ job_template: &job_template
         command:
           - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"
         args:
+          - "/home/prow/go/src/github.com/kyma-project/kyma/components/asset-store-controller-manager"
+        resources:
+          requests:
+            memory: 2.5Gi
+            cpu: 2
+
+old_job_template: &old_job_template
+  skip_report: false
+  decorate: true
+  path_alias: github.com/kyma-project/kyma
+  max_concurrency: 10
+  spec:
+    containers:
+      - image: eu.gcr.io/kyma-project/test-infra/buildpack-golang-kubebuilder:v20190208-813daef
+        securityContext:
+          privileged: true
+        command:
+          - "/home/prow/go/src/github.com/kyma-project/test-infra/prow/scripts/build.sh"
+        args:
           - "/home/prow/go/src/github.com/kyma-project/kyma/components/assetstore-controller-manager"
         resources:
           requests:
@@ -29,21 +48,21 @@ job_labels_template: &job_labels_template
 
 presubmits: # runs on PRs
   kyma-project/kyma:
-    - name: pre-master-kyma-components-assetstore-controller-manager
+    - name: pre-master-kyma-components-asset-store-controller-manager
       branches:
         - master
       <<: *job_template
-      run_if_changed: "^components/assetstore-controller-manager/"
+      run_if_changed: "^components/asset-store-controller-manager/"
       extra_refs:
       - <<: *test_infra_ref
         base_ref: master
       labels:
         <<: *job_labels_template
         preset-build-pr: "true"
-    - name: pre-rel07-kyma-components-assetstore-controller-manager
+    - name: pre-rel07-kyma-components-asset-store-controller-manager
       branches:
         - release-0.7
-      <<: *job_template
+      <<: *old_job_template
       always_run: true
       extra_refs:
       - <<: *test_infra_ref
@@ -51,10 +70,10 @@ presubmits: # runs on PRs
       labels:
         <<: *job_labels_template
         preset-build-release: "true"
-    - name: pre-rel08-kyma-components-assetstore-controller-manager
+    - name: pre-rel08-kyma-components-asset-store-controller-manager
       branches:
         - release-0.8
-      <<: *job_template
+      <<: *old_job_template
       always_run: true
       extra_refs:
       - <<: *test_infra_ref
@@ -65,11 +84,11 @@ presubmits: # runs on PRs
 
 postsubmits:
   kyma-project/kyma:
-    - name: post-master-kyma-components-assetstore-controller-manager
+    - name: post-master-kyma-components-asset-store-controller-manager
       branches:
         - master
       <<: *job_template
-      run_if_changed: "^components/assetstore-controller-manager/"
+      run_if_changed: "^components/asset-store-controller-manager/"
       extra_refs:
       - <<: *test_infra_ref
         base_ref: master
